LayoutTests/fast/viewport/viewport-31-expected.txt

ALERT: viewport size 980x700 scale 0.502857 with limits [0.502857, 5] and userScalable true
